make worker_monitor a separate file
