+++ lib/runtime/src/utils.rs
@@ -20,3 +20,4 @@ pub mod pool;
 pub mod stream;
 pub mod task;
 pub mod typed_prefix_watcher;
+pub mod worker_monitor;
